requirements.txt
Flask==3.1.2
langchain==1.1.3
langchain_core==1.2.0
langchain_openai==1.1.3
langgraph==1.0.5
pydantic==2.12.5
python-dotenv==1.2.1
tavily_python==0.7.15
